frontend: Fill in the home page
Fill in the home page with a simple card, welcoming the user and
offering that they join or create a room (in which they will later be
able to screen share to other users).

The desired room name can be entered into an input field, which is
already pre-filled with an autogenerated, random name, for the most
common use case where the user will be creating a new room. Otherwise,
they can enter the name of the room they want to join, or a custom name
with which to create a room.

Below the input field is a button that completes the action and joins
the room, however this is left unimplemented.

In the following commits, we should implemented the button to validate
the room name and then send the user to a room page, where they can
share their screen and view others' screen shares.