frontend: Create room page layout
Create the base layout for the room page, consisting of a title at the
top, a video stream region, and a series of buttons at the bottom.

Only the user button was added in this change, which will be used to
count the number of connected users, and later on to display their names
when clicked.

Some app-level changes were made to accomodate this new layout:
 - to achieve a full screen view, `h-screen` was added to the app div
 - a depencency on `@heroicons/vue` was added for button icons