frontend: room: Add connection status indicators
Add a status indicator (coloured circle) on the user button which
displays different colours depending on the connection state (which is
currently always disconnected). When a connection is actively being
established, show a ping animation on the status indicator to convey
this to the user.

Furthermore, keep a peer / user count - the number of other users that
we are connected to - and display this in the interface. When the data
channel to the server is established, add one to the user count to
indicate that the current user is fully connected and communicating with
the server. When the data channel is not established (and therefore the
user count is zero or otherwise inaccurate), change the colour of the
user counter to warning (usually yellow) to use as an additional data
channel status indicator.

With these UI elements, we now have full visibility into the WebRTC
connection status (which will be implemented next).