backend: Configure WebRTC
Add all of the necessary plumbing to start using WebRTC to connect
clients in the same room together.

In particular:
 - Add a dependency on webrtc
 - Add a dependency on clap
 - Introduce a command-line interface accepting options which are then
   used to configure the WebRTC public IP and ephemeral port range
 - Adjust the nix flake to include the same options as the command-line
   interface
 - Adjust the startup info logs to display the interfaces on which HTTP
   and WebRTC connections are being accepted

Note that we are forced to use an ephemeral port range due to a problem
with webrtc's UDP multiplexing (see #1). In the future we should resolve
this issue and use a single, fixed port.