frontend: Display 0 users when channel is closed
Previously, the users counter logic was to display the number of peers
registered so far, and add 1 if the current user is also connected (as
the server will not report the user's own connection to themselves).

However, if the user is not connected to the server, then there are 0
peers. To avoid an inaccurate counter being displayed to the user,
short-circuit any peer calculation logic and directly display users as
0.