backend: Refcount rooms
When sending create session / create stream messages to a room, instead
of relying on the room staying alive to process the message and retrying
otherwise, actively keep track of the number of clients which want to
keep the room alive through a new refcount mechanism.

Atomically exit when no references remain, allowing the next client to
create a new room if they were unable to get a reference to the room in
time.