backend: debug: Add option to emit webrtc-rs logs
Allow the backend to emit webrtc-rs logs to various sinks for debugging
and analysis when the `--webrtc-log-level` option is provided.

In particular, we use the first available sink in the following list:
 - a specific file descriptor, if specified via `--webrtc-log-fd`,
   allowing developers to redirect output to any file or terminal
 - journald, if the server is running as a systemd service
 - stdout

We also add some basic output colouring to imitate journald's colouring
(although the colour scheme is different).

Note that we intentionally do not expose the `--webrtc-log-fd`
configuration option in the nix flake as journald is currently
sufficiently flexible for our purposes. If we find that at some point in
the future we want to redirect logs away from journald, we can add the
relevant code and configuration.

Note also that file descriptors and journald are not available on
Windows, therefore a separate implementation exists which only logs to
stdout. We could add logging to files or other advanced logging schemes
in the future as required.