backend: Correctly serialize event payloads
Event payloads were serialized in snake_case rather than camelCase,
which means the stream IDs were being read as undefined on the client.

As a result, ending a stream which is not currently being displayed
would stop the stream which *is* being displayed, since both of their
IDs would be the same (undefined).