frontend: Switch to another stream on stream end
When the currently displayed stream ends, and there is more than one
stream in the map, display the first stream returned from the map in
place of the terminated stream.

With this logic we can now "stack" streams to some extent, even though
the interface only supports displaying a single stream at a time. When
multiple streams are started, only the last will be displayed, however
stopping the last stream will switch to another (e.g. the previously
started).

We could improve the logic to achieve true stacking by keeping a queue
of the order in which streams were added, however we should switch to a
multi-stream interface anyway, so this is not currently worth the
effort.