Inform users about codec incompatibility
Before starting a stream, detect which users are running clients which
support the stream codec. Only send the stream to those users; for the
rest, display an error message informing them what their current
supported codecs are and which one is missing specifically.

By applying this check, we avoid running into the scenario where a
user's browser does not support a codec yet we send them the stream
anyway. In this case, renegotiation over HTTP returns an error, meaning
the `negotiationneeded` callback exits early, the stream gets stuck in a
loading state, and the `negotiationneeded` event never fires again,
which means the user will no longer be kept in sync with the streams
being added and removed from the room.

Note that we only apply the lightest variant of compatibility checking:

 - First, we only check video codecs, since audio usually only uses opus
   which has wide support.

 - Second, we assume that if the client and server support the same
   codec, they will be able to negotiate a compatible format for
   transmission of the codec (via the `fmtp` attributes).

For that last point, I'm not entirely sure whether a scenario exists
where negotiation is not possible under the same codec, however if it
does, the client will fallback to the old behaviour described above,
where renegotiation fails. Transmitting with an unsupported audio codec,
including opus, will likely have the same effect, however there is no
current practical case where this happens.