frontend: Add an option to remember the room name
Allow the client to remember the last room that the user entered rather
than generating a new room name on every visit to the page.

Note that when using the main page to force an interaction before
entering a room, this option will not be available, to avoid messing
with the user's last remembered room name.