backend: Guarantee unique MSIDs
In the case that two clients coincidentally produce identical stream
IDs, prepend the streamfox-generated stream ID to the replicated track's
stream ID, thereby ensuring that MSIDs remain unique.

Also perform an identical operation to the track's ID for consistency.

In practice this happens if we launch two OBS instances with the same
settings, which causes errors in Chrome (which is more strict about
spec-accuracy than Firefox).

As per IETF RFC 8830 and the `token-char` grammar in RFC 8866, spaces
and slashes are not allowed in MSIDs, therefore use a plus to separate
the streamfox ID from the original track and stream ID.

As per RFC 8830, limit the MSID strings to 64 characters in length,
since although we expect the producer to already be following this
requirement, we may go over the limit while prepending additional data.
Since the strings are required to be ASCII by the RFCs,
`String::truncate` should not panic.