backend: Allow running WebRTC on a single port
Enable UDP multiplexing on a single port for WebRTC by specifying the
`--webrtc-port-mux` option (which is mutually exclusive with port ranges
specified with `--webrtc-port-min` / `--webrtc-port-max`).

UDP multiplexing is simpler as administrators only have to enable one
port in the firewall as opposed to a port range (which also reduces the
number of ports in use on the network), however port ranges are more
reliable in the case that the multiplexing code has bugs, therefore we
leave both options available.