frontend: Add custom video controls
Add and configure custom video controls via the media-chrome library.

Note that the styling is minimal and only what is required functionally;
in the future, we could significantly improve the style to make the
video player pretty and distinct.

Since media-chrome uses custom web elements, add a pattern match to
Vue's compiler options to inform Vue which custom elements are being
used and thereby skip the development warnings for missing components.

In addition, disable the lint `vue/no-deprecated-slot-attribute`, since
the custom web components that media-chrome provides were trigerring
this rule (and this error would not be useful for us anyway).