backend: Unwrap `data_channel` on `send`
Instead of checking if the data channel is present, apply a more
rigorous approach - unwrap and panic if there is no data channel,
thereby ensuring that the server logic is robust. If the data channel is
not present, we've made an error to be sending a message to the session
in the first place, and we should be aware of the error.