backend: Avoid redundant message param clones
Previously we had to clone the parameters when sending a message, as the
send could fail if the room had exited and we would have to try again
with the same parameters.

However, since we implemented the robust refcount mechanism, which
either succeeds or fails atomically, we are guaranteed to only send a
message once to a room that will not exit, and therefore we can move the
parameters into the message value as they are unused thereafter.