backend: Ensure video transceivers > 0
We currently do not support audio-only streams on both the frontend and
backend. On the frontend, we do not have a suitable user interface for
an audio-only stream (although the video interface may somewhat
suffice). On the backend, we perform a codec check against the incoming
video stream, and if no such stream is available, we crash during the
codec check.

As such, until we implement audio-only stream support (which is not
currently a desired use-case), we should prevent such stream connections
from succeeding and return an informative error.