frontend: Disable video browser context menu
Disable the browser-provided video context menu, unless explicitly
overridden by the user, as this menu shows controls such as play /
pause, which do not make sense for our use-case.

Since we now have all the necessary shortcuts in the control bar,
including picture-in-picture, we do not degrade the user experience with
this change.