backend: Add Unix domain socket support
Add support for HTTP requests over Unix domain sockets, which can be
used when the server is behind a reverse proxy to eliminate the TCP
overhead and thereby improve efficiency and throughput.

To facilitate this change on Nix systems, we now create a user and group
called `streamfox-live` and run the backend as this user, which enables
us to create a directory which hosts the socket that the backend has
permission to modify.