backend: Switch protocol naming order
When logging which sockets are listening on which addresses, log the
application-level protocol in brackets and the transport-level protocol
in the main text.

I believe it makes more sense to say that the transport protocol is
listening on an address because the data transport happens on that
address. The application protocols aren't involved in listening or
transporting data at all. I think it also makes more sense to show which
application protocols have which sockets open, rather than vice-versa,
because we think about and organise information on the application
level.

In addition, ordering the protocols this way makes adding logs for Unix
sockets in an upcoming change easier.